Avoid overflow in BackOffDelay and CombineDelay
Although we have MaxDelay option, that only kicks in after the
DelayTypeFunc calculation is done, and the calculation itself could
overflow.

For example, with initial delay of 1 second, the 34th retry would cause
the calculated back off delay to go negative, and the 35th retry will
cause it to go back to zero.

This is loosely ported from our implementation outside of the package:
https://github.com/reddit/baseplate.go/commit/660b6c5ff490eaf19c6ecc595762fe386d3ca5fc